+++ CHANGELOG.md
@@ -25,6 +25,7 @@ Please document your changes in this format:
 ## [8.7.2] - 2020-08-28
 ### Changed
 - Enable the Luxembourgish locale of Quasar and date-fns when this language is selected @dwaxweiler
+- Do not offer to join the playground group during signup, to avoid confusiong @tiltec
 
 ## [8.7.1] - 2020-08-27
 ### Added